requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
pillow==9.5.0
pytesseract==0.3.10
transformers==4.31.0
torch==2.0.1
